requirements/default.txt
astroid==2.3.1
bandit==1.6.2
Django==2.2.6
djangorestframework==3.10.3
gitdb2==2.0.6
GitPython==3.0.3
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pbr==5.4.3
pylint==2.4.2
pytz==2019.3
PyYAML==5.1.2
six==1.12.0
smmap2==2.0.5
sqlparse==0.3.0
stevedore==1.31.0
typed-ast==1.4.0
wrapt==1.11.2
